+++ repour/server/endpoint/validation.py
@@ -3,7 +3,6 @@ import logging
 from voluptuous import *
 
 from ...adjust import adjust as adjustmodule
-from ... import pull as pullmodule
 from ... import repo
 
 def mode_b_ify(raw):
@@ -72,52 +71,6 @@ adjust_modeb = Schema(
 
 external_to_internal = Schema(
     {"external_url": nonempty_str},
-    required=True,
-    extra=False,
-)
-
-#
-# Pull
-#
-pull_scm_raw = {
-    "name": name_str,
-    "type": Any(*pullmodule.scm_types),
-    Optional("ref"): nonempty_str,
-    "url": Url(), #pylint: disable=no-value-for-parameter
-    Optional("adjust"): bool,
-    Optional("callback"): callback_raw,
-}
-pull_scm = Schema(
-    pull_scm_raw,
-    required=True,
-    extra=False,
-)
-
-pull_archive_raw = {
-    "name": name_str,
-    "type": pullmodule.archive_type,
-    "url": Url(), #pylint: disable=no-value-for-parameter
-    Optional("adjust"): bool,
-    Optional("callback"): callback_raw,
-}
-pull_archive = Schema(
-    pull_archive_raw,
-    required=True,
-    extra=False,
-)
-
-pull_raw = Any(pull_scm, pull_archive)
-
-pull = Schema(
-    pull_raw,
-    required=True,
-    extra=False,
-)
-
-pull_scm_modeb = mode_b_ify(pull_scm_raw)
-pull_archive_modeb = mode_b_ify(pull_archive_raw)
-pull_modeb = Schema(
-    Any(pull_scm_modeb, pull_archive_modeb),
     required=True,
     extra=False,
 )